Update shard.yml
This changes the version to 0.2.0. Also add myself as a contributor and
increase the Crystal version to 0.33 (the version available at the
moment).

I'd like to eventually see the version number follow the release of
GLFW, but for now that can be done with the vX.Y branches (this one
under v3.3).